feat: 상품 상세 페이지 동적 데이터 노출 작업 완료
